+++ backend/src/main/java/org/camunda/optimize/dto/optimize/query/report/configuration/target_value/TargetDto.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public class TargetDto {
 
   private TargetValueUnit unit = TargetValueUnit.HOURS;
-  private Integer value = 2;
+  private Double value = 2.0;
 
   public TargetValueUnit getUnit() {
     return unit;
@@ -15,11 +15,11 @@ public class TargetDto {
     this.unit = unit;
   }
 
-  public Integer getValue() {
+  public Double getValue() {
     return value;
   }
 
-  public void setValue(Integer value) {
+  public void setValue(Double value) {
     this.value = value;
   }
